Reorganized Packet Handlers Before Calling It A Night
