--- src/main/java/me/alexisevelyn/crewmate/packethandler/packets/reliable/gamedata/HostSettingsPacket.java
@@ -1,4 +1,0 @@
-package me.alexisevelyn.crewmate.packethandler.packets.reliable.gamedata;
-
-public class HostSettingsPacket {
-}
